refactor(team): modularize cheat sheet helpers
